add validasi validation scheme

+++ adminkeu/pages/distribusi/distribusiread.php
@@ -25,13 +25,13 @@ if (isset($_SESSION['hasil'])) {
     if ($_SESSION['hasil_delete']) {
     ?>
         <div id='hasil_delete'></div>
-<?php }
+    <?php }
     unset($_SESSION['hasil_delete']);
 } elseif (isset($_SESSION['hasil_create'])) {
     if ($_SESSION['hasil_create']) {
     ?>
         <div id='hasil_create'></div>
-<?php }
+    <?php }
     unset($_SESSION['hasil_create']);
 } elseif (isset($_SESSION['hasil_update'])) {
     if ($_SESSION['hasil_update']) {
@@ -103,11 +103,12 @@ if (isset($_SESSION['hasil'])) {
                     $database = new Database;
                     $db = $database->getConnection();
 
-                    $selectsql = "SELECT a.*, d.*, k1.nama as supir, k1.upah_borongan usupir1, k2.nama helper1, k2.upah_borongan uhelper2, k3.nama helper2, k3.upah_borongan uhelper2, do1.nama distro1, do1.jarak jdistro1, do2.nama distro2, do2.jarak jdistro2, do3.nama distro3, do3.jarak jdistro3
+                    $selectsql = "SELECT a.*, d.*, k1.nama supir, k1.upah_borongan usupir1, k2.nama helper1, k2.upah_borongan uhelper2, k3.nama helper2, k3.upah_borongan uhelper2, v.nama validator, do1.nama distro1, do1.jarak jdistro1, do2.nama distro2, do2.jarak jdistro2, do3.nama distro3, do3.jarak jdistro3
                     FROM distribusi d INNER JOIN armada a on d.id_plat = a.id
                     LEFT JOIN karyawan k1 on d.driver = k1.id
                     LEFT JOIN karyawan k2 on d.helper_1 = k2.id
                     LEFT JOIN karyawan k3 on d.helper_2 = k3.id
+                    LEFT JOIN karyawan v on d.validasi_oleh = v.id
                     LEFT JOIN distributor do1 on d.nama_pel_1 = do1.id
                     LEFT JOIN distributor do2 on d.nama_pel_2 = do2.id
                     LEFT JOIN distributor do3 on d.nama_pel_3 = do3.id
@@ -123,15 +124,16 @@ if (isset($_SESSION['hasil'])) {
                         $distro1 = $row['distro1'] == NULL ? '-' : $row['distro1'];
                         $distro2 = $row['distro2'] == NULL ? '-' : $row['distro2'];
                         $distro3 = $row['distro3'] == NULL ? '-' : $row['distro3'];
-                        $bongkar = $row['bongkar'] == 0 ? 'TIDAK' : 'YA'; 
+                        $bongkar = $row['bongkar'] == 0 ? 'TIDAK' : 'YA';
                         $keterangan = $row['keterangan'] == NULL ? '-' : $row['keterangan'];
                         $jam_datang = $row['jam_datang'] == NULL ? '-' : date('d-m-Y H:i:s', strtotime($row['jam_datang']));
                         $tgl_validasi = $row['tgl_validasi'] == NULL ? '-' : date('d-m-Y H:i:s', strtotime($row['tgl_validasi']));
-                        $validasi_oleh = $row['validasi_oleh'] == NULL ? '-' : $row['validasi_oleh'];
+                        $validasi_oleh = $row['validator'] == NULL ? '-' : $row['validator'];
                         $estimasi_lama_perjalanan = date_diff(date_create($row['jam_berangkat']), date_create($row['estimasi_jam_datang']))->format('%d Hari %h Jam %i Menit %s Detik');
                         switch ($row['status']) {
                             case '0':
                                 $status = 'Belum Divalidasi';
+                                break;
                             case '1':
                                 $status = 'Divalidasi';
                                 break;
@@ -146,20 +148,20 @@ if (isset($_SESSION['hasil'])) {
                         <tr>
                             <td><?= $no++ ?></td>
                             <td><?= $row['no_perjalanan'] ?></td>
-                            <td><?= date('d-m-Y H:i:s',strtotime($row['tanggal'])) ?></td>
-                            <td><?= $row['plat'], ' - ' ,$row['jenis_mobil'];?></td>
+                            <td><?= date('d-m-Y H:i:s', strtotime($row['tanggal'])) ?></td>
+                            <td><?= $row['plat'], ' - ', $row['jenis_mobil']; ?></td>
                             <td><?= $supir ?></td>
                             <td><?= $helper1 ?></td>
                             <td><?= $helper2 ?></td>
-                            <td><?= $distro1?></td>
-                            <td><?= $distro2?></td>
-                            <td><?= $distro3?></td>
+                            <td><?= $distro1 ?></td>
+                            <td><?= $distro2 ?></td>
+                            <td><?= $distro3 ?></td>
                             <td><?= $bongkar ?></td>
-                            <td><?= $row['cup1']+$row['cup2']+$row['cup3'] ?></td>
-                            <td><?= $row['a3301']+$row['a3302']+$row['a3303'] ?></td>
-                            <td><?= $row['a5001']+$row['a5002']+$row['a5003'] ?></td>
-                            <td><?= $row['a6001']+$row['a6002']+$row['a6003'] ?></td>
-                            <td><?= $row['refill1']+$row['refill2']+$row['refill3'] ?></td>
+                            <td><?= $row['cup1'] + $row['cup2'] + $row['cup3'] ?></td>
+                            <td><?= $row['a3301'] + $row['a3302'] + $row['a3303'] ?></td>
+                            <td><?= $row['a5001'] + $row['a5002'] + $row['a5003'] ?></td>
+                            <td><?= $row['a6001'] + $row['a6002'] + $row['a6003'] ?></td>
+                            <td><?= $row['refill1'] + $row['refill2'] + $row['refill3'] ?></td>
                             <td><?= $row['jam_berangkat'] ?></td>
                             <td><?= $row['estimasi_jam_datang'] ?></td>
                             <td><?= $estimasi_lama_perjalanan ?></td>
@@ -169,12 +171,22 @@ if (isset($_SESSION['hasil'])) {
                             <td><?= $validasi_oleh ?></td>
                             <td><?= $status ?></td>
                             <td>
-                                <a href="?page=distribusiupdate&id=<?= $row['id']; ?>" class="btn btn-primary btn-sm mr-1">
-                                    <i class="fa fa-edit"></i> Ubah
-                                </a>
-                                <a href="?page=distribusidelete&id=<?= $row['id']; ?>" class="btn btn-danger btn-sm mr-1" id="deletedistribusi">
-                                    <i class="fa fa-trash"></i> Hapus
-                                </a>
+                                <?php if ($row['status'] == '0') { ?>
+                                    <a href="?page=distribusiupdate&id=<?= $row['id']; ?>" class="btn btn-primary btn-sm mr-1">
+                                        <i class="fa fa-edit"></i> Ubah
+                                    </a>
+                                    <a href="?page=distribusidelete&id=<?= $row['id']; ?>" class="btn btn-danger btn-sm mr-1" id="deletedistribusi">
+                                        <i class="fa fa-trash"></i> Hapus
+                                    </a>
+                                <?php } else if ($row['status'] == '1') { ?>
+                                    <a href="#" class="btn btn-primary btn-sm mr-1 disabled" role="button" aria-disabled="true" id="distribusidisable">
+                                        <i class="fa fa-edit"></i> Ubah
+                                    </a>
+                                    <a href="#" class="btn btn-danger btn-sm mr-1 disabled" role="button" aria-disabled="true" id="distribusidisable">
+                                        <i class="fa fa-trash"></i> Hapus
+                                    </a>
+                                <?php }; ?>
+
                             </td>
                         </tr>
                     <?php } ?>
@@ -208,6 +220,13 @@ include_once "../partials/scriptdatatables.php";
                 }
             })
         });
+        // $('a#distribusidisable').click(function() {
+        //     Swal.fire({
+        //         icon: 'error',
+        //         title: 'Ups',
+        //         text: 'Data yang sudah divalidasi tidak bisa diubah!',
+        //     })
+        // });
         if ($('div#hasil_delete').length) {
             Swal.fire({
                 title: 'Deleted!',
